+++ x/emissions/module/worker_rewards.go
@@ -15,43 +15,59 @@ func GetWorkersRewardsInferenceTask(
 	preward float64,
 	totalRewardsInferenceTask float64,
 ) ([]float64, error) {
-	// Get Network Losses from the last block
-	// TODO: Change this after merging with kenny's PR
-	networkLosses, err := am.keeper.GetLossBundles(ctx, topicId, block)
+	// Get network loss
+	networkLosses, err := am.keeper.GetNetworkLossBundleAtOrBeforeBlock(ctx, topicId, block)
 	if err != nil {
 		return nil, err
 	}
 
-	// Get Score for each worker
-	// TODO: Sort by block (?)
-	var scores [][]float64
-	for _, networkLosses := range networkLosses.LossBundles {
-
-		// TODO: Sort by worker (?)
-		var workersScores []float64
-		for _, oneOutLoss := range networkLosses.OneOutLosses {
-			workerScore := GetWorkerScore(float64(networkLosses.CombinedLoss.BigInt().Int64()), float64(oneOutLoss.Value.BigInt().Int64()))
-
-			workerAddr, err := sdk.AccAddressFromBech32(oneOutLoss.Worker)
-			if err != nil {
-				return nil, err
-			}
-
-			// Persist worker score
-			am.keeper.InsertWorkerInferenceScore(ctx, topicId, block, types.Score{
-				TopicId:     topicId,
-				BlockNumber: block,
-				Address:     workerAddr.String(),
-				Score:       workerScore,
-			})
-
-			workersScores = append(workersScores, workerScore)
+	// Get new score for each worker
+	// TODO: Sort by worker (?)
+	var scores [][]types.Score
+	for _, oneOutLoss := range networkLosses.OneOutLosses {
+		workerAddr, err := sdk.AccAddressFromBech32(oneOutLoss.Worker)
+		if err != nil {
+			return nil, err
 		}
-		scores = append(scores, workersScores)
+
+		// Get worker last scores
+		workerLastScores, err := am.keeper.GetLastWorkerInferenceScores(ctx, topicId, block, workerAddr)
+		if err != nil {
+			return nil, err
+		}
+
+		// Calculate new score
+		workerNewScore := GetWorkerScore(float64(networkLosses.CombinedLoss.BigInt().Int64()), float64(oneOutLoss.Value.BigInt().Int64()))
+
+		scoreToAdd := types.Score{
+			TopicId:     topicId,
+			BlockNumber: block,
+			Address:     workerAddr.String(),
+			Score:       workerNewScore,
+		}
+
+		// Persist worker score
+		am.keeper.InsertWorkerInferenceScore(ctx, topicId, block, scoreToAdd)
+
+		// Add new score in the last scores array
+		workerLastScores = append(workerLastScores, scoreToAdd)
+
+		// Add worker score in the scores array
+		scores = append(scores, workerLastScores)
+	}
+
+	// Convert scores to float64
+	var scoresFloat64 [][]float64
+	for _, workerScoreArray := range scores {
+		var scoresOneOutFloat64 []float64
+		for _, score := range workerScoreArray {
+			scoresOneOutFloat64 = append(scoresOneOutFloat64, score.Score)
+		}
+		scoresFloat64 = append(scoresFloat64, scoresOneOutFloat64)
 	}
 
 	// Get Worker Reward Fractions
-	workersRewardsFractions, err := GetWorkerRewardFractions(scores, preward)
+	workersRewardsFractions, err := GetWorkerRewardFractions(scoresFloat64, preward)
 	if err != nil {
 		return nil, err
 	}
@@ -66,7 +82,6 @@ func GetWorkersRewardsInferenceTask(
 	return workerRewards, nil
 }
 
-// TODO: Add litepaper references
 func GetWorkersRewardsForecastTask(
 	ctx sdk.Context,
 	am AppModule,
@@ -75,58 +90,74 @@ func GetWorkersRewardsForecastTask(
 	preward float64,
 	totalRewardsForecastTask float64,
 ) ([]float64, error) {
-	// Get Network Losses from the last block
-	// TODO: Change this after merging with kenny's PR
-	networkLosses, err := am.keeper.GetLossBundles(ctx, topicId, block)
+	// Get network loss
+	networkLosses, err := am.keeper.GetNetworkLossBundleAtOrBeforeBlock(ctx, topicId, block)
 	if err != nil {
 		return nil, err
 	}
 
-	// Get Score for each worker
-	// TODO: Sort by block (?)
-	var scores [][]float64
-	for _, networkLosses := range networkLosses.LossBundles {
+	// Get worker scores for one out loss
+	var workersScoresOneOut []float64
+	for _, oneOutLoss := range networkLosses.OneOutLosses {
+		workerScore := GetWorkerScore(float64(networkLosses.CombinedLoss.BigInt().Int64()), float64(oneOutLoss.Value.BigInt().Int64()))
+		workersScoresOneOut = append(workersScoresOneOut, workerScore)
+	}
 
-		// TODO: Sort by worker (?)
-		var workersScoresOneOut []float64
-		for _, oneOutLoss := range networkLosses.OneOutLosses {
-			workerScore := GetWorkerScore(float64(networkLosses.CombinedLoss.BigInt().Int64()), float64(oneOutLoss.Value.BigInt().Int64()))
-			workersScoresOneOut = append(workersScoresOneOut, workerScore)
+	// Get worker scores for one in naive loss
+	var workersScoresOneIn []float64
+	for _, oneInNaiveLoss := range networkLosses.OneInNaiveLosses {
+		workerScore := GetWorkerScore(float64(networkLosses.NaiveLoss.BigInt().Int64()), float64(oneInNaiveLoss.Value.BigInt().Int64()))
+		workersScoresOneIn = append(workersScoresOneIn, workerScore)
+	}
+
+	// Get new score for each worker
+	numForecasters := len(workersScoresOneOut)
+	fUniqueAgg := GetfUniqueAgg(float64(numForecasters))
+	var scores [][]types.Score
+	for i, workerScoreOneOut := range workersScoresOneOut {
+		workerAddr, err := sdk.AccAddressFromBech32(networkLosses.OneInNaiveLosses[i].Worker)
+		if err != nil {
+			return nil, err
 		}
 
-		var workersScoresOneIn []float64
-		for _, oneInNaiveLoss := range networkLosses.OneInNaiveLosses {
-			workerScore := GetWorkerScore(float64(networkLosses.NaiveLoss.BigInt().Int64()), float64(oneInNaiveLoss.Value.BigInt().Int64()))
-			workersScoresOneIn = append(workersScoresOneIn, workerScore)
+		// Get worker last scores
+		workerLastScores, err := am.keeper.GetLastWorkerForecastScores(ctx, topicId, block, workerAddr)
+		if err != nil {
+			return nil, err
 		}
 
-		var workersFinalScoresTimestep []float64
-		numForecasters := len(workersScoresOneOut)
-		fUniqueAgg := GetfUniqueAgg(float64(numForecasters))
-		for i, workerScoreOneOut := range workersScoresOneOut {
-			workerFinalScore := GetFinalWorkerScoreForecastTask(workersScoresOneIn[i], workerScoreOneOut, fUniqueAgg)
+		// Calculate new score	
+		workerFinalScore := GetFinalWorkerScoreForecastTask(workersScoresOneIn[i], workerScoreOneOut, fUniqueAgg)
 
-			workerAddr, err := sdk.AccAddressFromBech32(networkLosses.OneInNaiveLosses[i].Worker)
-			if err != nil {
-				return nil, err
-			}
-
-			// Persist worker score
-			am.keeper.InsertWorkerForecastScore(ctx, topicId, block, types.Score{
-				TopicId:     topicId,
-				BlockNumber: block,
-				Address:     workerAddr.String(),
-				Score:       workerFinalScore,
-			})
-
-			workersFinalScoresTimestep = append(workersFinalScoresTimestep, workerFinalScore)
+		scoreToAdd := types.Score{
+			TopicId:     topicId,
+			BlockNumber: block,
+			Address:     workerAddr.String(),
+			Score:       workerFinalScore,
 		}
 
-		scores = append(scores, workersFinalScoresTimestep)
+		// Persist worker score
+		am.keeper.InsertWorkerForecastScore(ctx, topicId, block, scoreToAdd)
+
+		// Add new score in the last scores array
+		workerLastScores = append(workerLastScores, scoreToAdd)
+
+		// Add worker score in the scores array
+		scores = append(scores, workerLastScores)
+	}
+
+	// Convert scores to float64
+	var scoresFloat64 [][]float64
+	for _, workerScoreArray := range scores {
+		var scoresOneOutFloat64 []float64
+		for _, score := range workerScoreArray {
+			scoresOneOutFloat64 = append(scoresOneOutFloat64, score.Score)
+		}
+		scoresFloat64 = append(scoresFloat64, scoresOneOutFloat64)
 	}
 
 	// Get Worker Reward Fractions
-	workersRewardsFractions, err := GetWorkerRewardFractions(scores, preward)
+	workersRewardsFractions, err := GetWorkerRewardFractions(scoresFloat64, preward)
 	if err != nil {
 		return nil, err
 	}
